python/requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
google-genai
imagekitio
mysql-connector-python
apscheduler
